Moved DTOs to ViewModels folder.
